requirements.txt
# ============================================
# Multimodal AI Production - Dependencies
# ============================================

# Core ML Framework
torch>=2.4.0
torchvision>=0.19.0
torchaudio>=2.4.0

# Hugging Face Ecosystem
transformers>=4.45.0
accelerate>=0.34.0
datasets>=3.0.0
peft>=0.13.0
trl>=0.11.0
bitsandbytes>=0.44.0
tokenizers>=0.20.0
safetensors>=0.4.0
huggingface-hub>=0.25.0

# Unsloth - Memory efficient training
unsloth[colab-new]>=2024.11

# Audio Processing — TODO-8: faster-whisper preferred, openai-whisper fallback
faster-whisper>=1.0.0
openai-whisper>=20240930
soundfile>=0.12.0
librosa>=0.10.0

# Document Parsing — TODO-7: PDF/DOCX support
PyMuPDF>=1.24.0
pdfplumber>=0.11.0
python-docx>=1.1.0

# Web Search
duckduckgo-search>=7.0.0

# Report Generation (PDF)
reportlab>=4.0.0

# Image Processing
Pillow>=10.0.0
opencv-python-headless>=4.8.0

# RAG - Vector Database & Embeddings
sentence-transformers>=3.0.0
pgvector>=0.3.0
psycopg2-binary>=2.9.0
supabase>=2.0.0
chromadb>=0.5.0
faiss-cpu>=1.8.0

# API Server
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
pydantic>=2.9.0
pydantic-settings>=2.5.0

# Inference Optimization
vllm>=0.6.0;sys_platform=="linux"
onnxruntime>=1.19.0

# Monitoring & Logging
loguru>=0.7.0
wandb>=0.18.0
rich>=13.0.0

# Configuration
pyyaml>=6.0.0
python-dotenv>=1.0.0

# Utilities
numpy>=1.26.0
pandas>=2.2.0
tqdm>=4.66.0
httpx>=0.27.0
aiofiles>=24.0.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.24.0
